% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA.r
\name{EDMA}
\alias{EDMA}
\title{Compute Pairwise Landmark Distances}
\usage{
EDMA(A)
}
\arguments{
\item{A}{k x m x n array of landmarks}
}
\value{
returns a matrix with each row containing the pairwise distances between landmarks
}
\description{
Compute pairwise landmark distances from landmark array
}
