% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILDS.r
\name{ILDSR2}
\alias{ILDSR2}
\title{Compute R2 for Interlandmark Distances}
\usage{
ILDSR2(
  x,
  groups,
  R2tol = 0.95,
  bg.rounds = 999,
  wg.rounds = 999,
  which = NULL,
  reference = NULL,
  target = NULL,
  mc.cores = 1,
  plot = FALSE,
  silent = FALSE
)
}
\arguments{
\item{x}{array containing landmarks}

\item{groups}{vector containing group assignments. If more than two levels, a pair of needs to be specified.}

\item{R2tol}{numeric: upper percentile for SILD R2 in relation to factor}

\item{bg.rounds}{numeric: number of permutation rounds to assess between group differences}

\item{wg.rounds}{numeric: number of rounds to assess noise within groups by bootstrapping.}

\item{which}{integer (optional): in case the factor levels are > 2 this determins which factorlevels to use}

\item{reference}{matrix containing start config landmarks. If NULL, it will be computed as mean for group 1.}

\item{target}{matrix containing target config landmarks. If NULL, it will be computed as mean for group 2.}

\item{mc.cores}{integer: number of cores to use for permutation tests.}

\item{plot}{logical: if TRUE show graphical output of steps involved}

\item{silent}{logical: suppress console output}
}
\value{
A list containing:
\item{largeR2}{containing landmark information with the highest R2 (Andrea please specify here)}
\item{allR2}{vector with ILD specific R2-values, sorted decreasingly}
\item{reftarILDS}{matrix with columns containing ILDs for reference and target shapes}
\item{sampleILD}{matrix containing ILDs of entire sample}
\item{R2tol}{R2-threshold used}
}
\description{
Compute R2 for Interlandmark Distances explaining between group differences
}
\examples{
require(Morpho)
data(boneData)
proc <- procSym(boneLM)
groups <- name2factor(boneLM,which=3)
ilds <- ILDSR2(proc$rotated,groups,plot=TRUE,wg.rounds=999,mc.cores=2)
}
