% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILDS.r
\name{plot.ILDSR2}
\alias{plot.ILDSR2}
\title{Plot the ILDS with the relevant ILDS ighlighted}
\usage{
\method{plot}{ILDSR2}(x, ref = TRUE, relcol = "red", rescol = "black", ...)
}
\arguments{
\item{x}{output of function \code{\link{ILDSR2}}}

\item{ref}{logical: if TRUE, the reference shape defined in  \code{\link{ILDSR2}} will be plotted. Otherwise the target is used.}

\item{relcol}{color of relevant ILDs}

\item{rescol}{color of "irrelevant" ILDs}

\item{...}{additional parameters passed to  \code{\link{deformGrid2d}} /  \code{\link{deformGrid3d}}.}
}
\description{
Plot the ILDS with the relevant ILDS ighlighted
}
\examples{
## 3D Example
require(Morpho)
data(boneData)
proc <- procSym(boneLM)
groups <- name2factor(boneLM,which=3)
ilds <- ILDSR2(proc$rotated,groups,plot=FALSE,bg.rounds=0,wg.rounds=0)
plot(ilds)

## 2D Example
require(shapes)
gor.dat <- bindArr(gorf.dat,gorm.dat,along=3)
sex <- factor(c(rep("f",30),rep("m",29)))
procg <- procSym(gor.dat)
ildsg <- ILDSR2(procg$rotated,sex,plot=FALSE,bg.rounds=0,wg.rounds=0)
plot(ildsg)
}
