% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA.r
\name{ILDSR2}
\alias{ILDSR2}
\title{Compute R2 for Interlandmark Distances}
\usage{
ILDSR2(x, groups, startref, target, R2tol = 0.95, plot = FALSE)
}
\arguments{
\item{x}{array containing landmarks}

\item{groups}{vector containing group assignments}

\item{startref}{matrix containing start config landmarks}

\item{target}{matrix containing target config landmarks}

\item{R2tol}{numeric: upper percentile for SILD R2 in relation to factor}

\item{plot}{logical: if TRUE show graphical output of steps involved}
}
\value{
matrix containing landmark information with the highest R2 (Andrea please specify here)
}
\description{
Compute R2 for Interlandmark Distances explaining between group differences
}
\examples{
require(Morpho)
data(boneData)
proc <- procSym(boneLM)
groups <- name2factor(boneLM,which=3)
startref <- arrMean3(proc$rotated[,,groups=="ch"])
target <- arrMean3(proc$rotated[,,groups=="eu"])
ilds <- ILDSR2(proc$rotated,groups,startref,target,plot=TRUE)
}
