% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA.r
\name{plot.ILDSR2}
\alias{plot.ILDSR2}
\title{Plot the ILDS with the relevant ILDS ighlighted}
\usage{
\method{plot}{ILDSR2}(x, ref = TRUE, ...)
}
\arguments{
\item{x}{output of function \code{\link{ILDSR2}}}

\item{ref}{logical: if TRUE, the reference shape defined in  \code{\link{ILDSR2}} will be plotted. Otherwise the target is used.}

\item{...}{additional parametr - currently not used.}
}
\description{
Plot the ILDS with the relevant ILDS ighlighted
}
\examples{
require(Morpho)
data(boneData)
proc <- procSym(boneLM)
groups <- name2factor(boneLM,which=3)
reference <- arrMean3(proc$rotated[,,groups=="ch"])
target <- arrMean3(proc$rotated[,,groups=="eu"])
ilds <- ILDSR2(proc$rotated,groups,reference,target,plot=TRUE)
plot(ilds)
}
